improving_Search/requirements2.txt
fastapi
uvicorn
pymongo
python-dotenv
sentence-transformers
torch
rapidfuzz
requests